feat: 增加entity value type 返回

+++ iab-integrations/msc-integration/src/main/java/com/milesight/iab/integration/msc/service/MscDeviceService.java
@@ -170,7 +170,7 @@ public class MscDeviceService {
         entities.add(new EntityBuilder()
                 .identifier(MscIntegrationConstants.InternalPropertyKey.LAST_SYNC_TIME)
                 .property(MscIntegrationConstants.InternalPropertyKey.LAST_SYNC_TIME, AccessMod.R)
-                .valueType(EntityValueType.INT)
+                .valueType(EntityValueType.LONG)
                 .build());
     }
